Add playtest page to SSR as well.
