+++ src/components/DndProvider.js
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { DndProvider } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
 import TouchBackend from 'react-dnd-touch-backend';
