+++ src/cube_draft.js
@@ -2,9 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import DraftView from './components/DraftView';
+import ErrorBoundary from './components/ErrorBoundary';
 
 const wrapper = document.getElementById('react-root');
-const element = <DraftView {...reactProps} />;
+const element = <ErrorBoundary className="mt-3"><DraftView {...reactProps} /></ErrorBoundary>;
 if (wrapper) {
   if (wrapper.children.length === 0) {
     ReactDOM.render(element, wrapper);
