+++ views/cube/cube_blog.pug
@@ -61,3 +61,4 @@ block scripts
   include ../react
   script(src='/js/blog.js')
   script(src='/js/cube_blog.bundle.js')
+  script(src='/js/texteditor.js')
