+++ views/cube/cube_playtest.pug
@@ -15,10 +15,6 @@ block cube_toolbar
 
 block cube_content
   include ../dynamic_flash
-  input#deckInput(type='hidden', value=decks)
-  input#cubeID(type='hidden', value=cube_id)
-  input#canEdit(type='hidden', value=user && user.id == cube.owner)
-  input#draftFormats(type='hidden', value=cube.draft_formats)
   #react-root
   if user
     if user.id == cube.owner
@@ -78,5 +74,8 @@ block cube_content
 
 block scripts
   include ../react
+  script(type='text/javascript').
+    var reactProps = !{JSON.stringify(reactProps)};
   script(src='/js/cube_playtest.bundle.js')
-  script(src='/js/packcraft.js')
+  script(src='/js/packcraft.js', async)
+  script(src='/js/texteditor.js', async)
