+++ views/texteditor.pug
@@ -1,6 +1,4 @@
-
 link(rel='stylesheet' href='/css/texteditor.css')
-
 
 br
 em To tag cards in post, use '[[cardname]]'. E.g. [[Island]]
@@ -21,7 +19,3 @@ em To tag cards in post, use '[[cardname]]'. E.g. [[Island]]
         a.toolbar-item(href='#' data-command='insertUnorderedList') ul
         a.toolbar-item(href='#' data-command='insertOrderedList') ol
   #editor(contenteditable='') !{editorvalue}
-    
-    
-    
-script(src='/js/texteditor.js')
